esmify, import model pkg properly
