+++ engine/components/constants.js
@@ -1,3 +1,3 @@
-module.exports = {
+export default {
   optionalText: ' (Optional)'
 }
