+++ engine/package.json
@@ -8,7 +8,7 @@
     "fix-lint": "yarn run eslint ./ -c ../.eslintrc.json --fix",
     "lint": "yarn run eslint ./ -c ../.eslintrc.json",
     "build": "yarn run babel --root-mode upward src -d lib --source-maps inline",
-    "watch": "bash -c 'nohup yarn run babel src --root-mode upward -d lib --source-maps inline --watch'",
+    "watch": "yarn build --watch",
     "test-lab": "lab -T test/.transform.js -S -v -P 'test' test/cases -r console -o stdout -r html -o unit-test.html -I version,regeneratorRuntime",
     "test": "yarn flow && yarn lint && yarn test-lab",
     "test-lab-cov": "lab -T test/.transform.js test/**/*.test.js -S -v -r console -o stdout -r html -o test-coverage/lab/unit-test.html -r junit -o test-results/lab/unit-test.xml -I version,regeneratorRuntime",
